[RESTEASY-1862] Move JettyClient engine to separate module
