+++ pom.xml
@@ -102,6 +102,7 @@
                 <module>resteasy-jaxrs</module>
                 <module>resteasy-jaxrs-services</module>
                 <module>resteasy-client</module>
+                <module>resteasy-client-jetty</module>
                 <module>resteasy-legacy</module>
                 <module>providers</module>
                 <module>resteasy-bom</module>
@@ -154,6 +155,7 @@
                 <module>tjws</module>
                 <module>resteasy-jaxrs</module>
                 <module>resteasy-client</module>
+                <module>resteasy-client-jetty</module>
                 <module>providers</module>
                 <module>resteasy-bom</module>
                 <module>resteasy-cache</module>
