+++ resteasy-client-jetty/src/main/java/org/jboss/resteasy/client/jaxrs/engines/jetty/JettyClientResponse.java
@@ -1,4 +1,4 @@
-package org.jboss.resteasy.client.jaxrs.engines;
+package org.jboss.resteasy.client.jaxrs.engines.jetty;
 
 import java.io.IOException;
 import java.io.InputStream;
